fix: resolve code quality issues in migrations and scripts
- Migration: Change view_count, search_frequency, market_mentions from INT to BIGINT to match Prisma schema and prevent data truncation
- CSV export: Add checks for newlines and quotes to properly escape fields per RFC 4180 standard
- Career stage: Replace hardcoded year 2026 with dynamic new Date().getFullYear() for longevity
- Error handling: Add error logging to fetchJSON function for better API failure visibility

+++ prisma/migrations/003_artist_metrics_trending/migration.sql
@@ -10,10 +10,10 @@ ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS last_viewed TIMESTAMP;
 
 -- Step 1: Add new columns
 ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS metric_window VARCHAR(10) NOT NULL DEFAULT '7d';
-ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS view_count INT NOT NULL DEFAULT 0;
-ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS search_frequency INT NOT NULL DEFAULT 0;
-ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS market_mentions INT NOT NULL DEFAULT 0;
-ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS trending_rank INT;
+ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS view_count BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS search_frequency BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS market_mentions BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS trending_rank BIGINT;
 ALTER TABLE artist_metrics ADD COLUMN IF NOT EXISTS computed_at TIMESTAMP NOT NULL DEFAULT now();
 
 -- Step 2: Backfill view_count and search_frequency from existing columns
